+++ .pre-commit-config.yaml
@@ -31,7 +31,7 @@ repos:
     hooks:
       - id: mypy
         exclude: examples/simple/setup.py
-        additional_dependencies: [types-requests]
+        additional_dependencies: [types-requests, types-PyYAML]
 
   - repo: https://github.com/pre-commit/mirrors-prettier
     rev: v2.6.2
