+++ docs/index.rst
@@ -43,6 +43,7 @@ Jupyter's Events library can be installed from PyPI.
    pages/application
    pages/schemas
    pages/redaction_policies
+   pages/demo
 
 Indices and tables
 ------------------
