+++ docs/pages/demo.md
@@ -1,0 +1,8 @@
+# Jupyter Events Demo
+
+```{retrolite} demo-notebook.ipynb
+---
+width: 100%
+height: 600px
+---
+```
